+++ src/scorer/feed/diversity_feed_scorer.ts
@@ -33,11 +33,11 @@ export default class DiversityFeedScorer extends FeedScorer {
 
         // Count how many times each account and each trending tag are seen in the feed
         sortedToots.forEach((toot) => {
-            toot.withRetoot().forEach((t) => accountsInFeed.incrementCount(t.account.webfingerURI));
+            toot.withRetoot.forEach((t) => accountsInFeed.incrementCount(t.account.webfingerURI));
 
             // Penalties for trending tags are similar to those for accounts but we base the max penalty
             // on the TrendingTag's numAccounts property (the fediverse-wide number of accounts using that tag)
-            toot.realToot().trendingTags!.forEach((tag) => {
+            toot.realToot.trendingTags!.forEach((tag) => {
                 const penalizedTag = trendingTagsInFeed.incrementCount(tag.name);
                 penalizedTag.numAccounts = Math.max(tag.numAccounts || 0, penalizedTag.numAccounts || 0);
                 penalizedTag.penaltyIncrement = penalizedTag.numAccounts / penalizedTag.numToots!;
@@ -51,13 +51,13 @@ export default class DiversityFeedScorer extends FeedScorer {
         // The biggest penalties are applied to toots encountered first. We want to penalize the oldest toots the most.
         return sortedToots.reduce(
             (tootScores, toot) => {
-                toot.withRetoot().forEach((t) => {
+                toot.withRetoot.forEach((t) => {
                     const penalty = this.computePenalty(accountsInFeed, t.account.webfingerURI);
                     incrementCount(tootScores, toot.uri, penalty);
                 });
 
                 // Additional penalties for trending tags
-                (toot.realToot().trendingTags || []).forEach((tag) => {
+                (toot.realToot.trendingTags || []).forEach((tag) => {
                     const penalty = this.computePenalty(trendingTagsInFeed, tag.name);
 
                     // Don't apply penalty to followed accounts/tags